Schema - Restructure FieldTypeIntegration component folder layout (#4103)
## Summary

Pure refactor of `src/shell/components/FieldTypeIntegration/` to make
the folder layout self-describing and shrink consumer import paths. **No
behavior changes.**

### What moved
- `IntegrationFieldConfigure/` → `Configure/`
- `IntegrationFieldSelect/` → `Select/`
- `components/` → `Shared/`

### What split
- `components/Wrappers.tsx` → `Shared/FieldWrapper.tsx` +
`Shared/FormWrapper.tsx`
- `configs.ts` → `constants.ts` (display-type catalogs, color map,
display-options config) + `types.ts` (TypeScript shapes only)

### What's new
- Barrel `src/shell/components/FieldTypeIntegration/index.tsx` exporting
`IntegrationFieldConfigure` and `IntegrationFieldSelect` so consumers
can import via `shell/components/FieldTypeIntegration` instead of
reaching into a subpath.

### Consumers updated
- `src/apps/content-editor/src/app/components/Editor/Field/Field.tsx`
- `src/apps/schema/src/app/components/AddFieldModal/FieldFormInput.tsx`

Both now use the shorter barrel import.

---------
