+++ src/apps/content-editor/src/app/components/Editor/Field/Field.tsx
@@ -52,7 +52,7 @@ import {
 import { FieldTypeMedia } from "../../FieldTypeMedia";
 import { debounce, parseInt } from "lodash";
 import { useRegisterRef } from "../../../../../../../engine/useRegisterRef";
-import IntegrationFieldSelect from "../../../../../../../shell/components/FieldTypeIntegration/IntegrationFieldSelect";
+import { IntegrationFieldSelect } from "shell/components/FieldTypeIntegration";
 import { useDebouncedInput } from "../../../../../../../shell/hooks/useDebouncedInput";
 import { format as fmt } from "date-fns";
 import { FieldTypeRepeater } from "shell/components/FieldTypeRepeater";
