+++ src/shell/components/FieldTypeIntegration/Configure/ConnectToApi.tsx
@@ -25,7 +25,8 @@ import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
 import AutorenewRoundedIcon from "@mui/icons-material/AutorenewRounded";
 import AddCircleIcon from "@mui/icons-material/AddCircle";
 import CloseOutlinedIcon from "@mui/icons-material/CloseOutlined";
-import { FormWrapper, FieldWrapper } from "../components/Wrappers";
+import { FormWrapper } from "../Shared/FormWrapper";
+import { FieldWrapper } from "../Shared/FieldWrapper";
 import { IntegrationRequestHeaders } from "../../../services/types";
 import { validateUrl } from "../utils";
 import useIntegrationField from "../useIntegrationField";
