+++ src/shell/components/FieldTypeIntegration/Shared/FieldWrapper.tsx
@@ -1,7 +1,6 @@
-import { Box, Paper, Tooltip, Typography } from "@mui/material";
+import { Box, Tooltip, Typography } from "@mui/material";
 
 import InfoIcon from "@mui/icons-material/Info";
-import { ReactNode } from "react";
 
 export type FieldWrapperProps = {
   name?: string;
@@ -11,12 +10,6 @@ export type FieldWrapperProps = {
   isRequired?: boolean;
   error?: string;
   children: React.ReactNode;
-};
-
-export type FormWrapperProps = {
-  width: string | number;
-  height: string | number;
-  children: ReactNode;
 };
 
 export const FieldWrapper = ({
@@ -87,26 +80,5 @@ export const FieldWrapper = ({
         </Typography>
       )}
     </Box>
-  );
-};
-
-export const FormWrapper = ({ width, height, children }: FormWrapperProps) => {
-  return (
-    <Paper
-      sx={{
-        width: width,
-        height: height,
-        maxHeight: "calc(100vh - 40px)",
-        borderRadius: 2,
-        position: "relative",
-        boxSizing: "border-box",
-        display: "flex",
-        flexDirection: "column",
-        justifyContent: "space-between",
-        alignItems: "stretch",
-      }}
-    >
-      {children}
-    </Paper>
   );
 };
