+++ src/shell/components/FieldTypeIntegration/index.tsx
@@ -1,0 +1,2 @@
+export { default as IntegrationFieldConfigure } from "./Configure";
+export { default as IntegrationFieldSelect } from "./Select";
